Add cache timeouts

+++ R/cancensus.R
@@ -293,15 +293,22 @@ cancensus.list_datasets <- function(use_cache = TRUE) {
     message("Querying CensusMapper API for available datasets...")
     response <- httr::GET("https://censusmapper.ca/api/v1/list_datasets",
                           httr::accept_json())
-    cancensus.handle_status_code(response, cache_file)
+    cancensus.handle_status_code(response, NULL)
     result <- httr::content(response, type = "text", encoding = "UTF-8") %>%
       jsonlite::fromJSON() %>%
       dplyr::as_data_frame()
+    attr(result, "last_updated") <- Sys.time()
     save(result, file = cache_file)
     result
   } else {
     message("Reading dataset list from local cache.")
     load(file = cache_file)
+    last_updated <- attr(result, "last_updated")
+    if (is.null(last_updated) ||
+          difftime(Sys.time(), last_updated, units = "days") > 1) {
+      warning(paste("Cached dataset list may be out of date. Set `use_cache =",
+                    "FALSE` to update it."))
+    }
     result
   }
 }
@@ -310,31 +317,30 @@ cancensus.list_datasets <- function(use_cache = TRUE) {
 #'
 #' @param dataset The dataset to query for available vectors, e.g.
 #'   \code{"CA16"}.
-#' @param use_cache If set to TRUE (the default) data will be read from the local cache if available.
+#' @param use_cache If set to TRUE (the default) data will be read from a local
+#'   cache, if available. If set to FALSE, query the API for the data, and
+#'   refresh the local cache with the result.
 #'
 #' @export
 #'
 #' @importFrom dplyr %>%
 cancensus.list_vectors <- function(dataset, use_cache=TRUE) {
   # TODO: Validate dataset?
-  result <- NULL
-  dir.create('data_cache', showWarnings = FALSE) # make sure cache directory exists
-  vector_file=paste0("data_cache/vector_list_",dataset,".csv")
-  if (!use_cache || !file.exists(vector_file)) {
-    url <- paste0("https://censusmapper.ca/api/v1/vector_info/", dataset, ".csv")
-    response <- httr::GET(url, httr::write_disk(vector_file, overwrite = TRUE),
+  cache_dir <- system.file("cache/", package = "cancensus")
+  cache_file <- paste0(cache_dir, dataset, "_vectors.rda")
+  if (!use_cache || !file.exists(cache_file)) {
+    message("Querying CensusMapper API for vectors data...")
+    response <- httr::GET(paste0("https://censusmapper.ca/api/v1/vector_info/",
+                                 dataset, ".csv"),
                           httr::progress())
-    cancensus.handle_status_code(response,vector_file)
-  } else {
-    message("Reading vector information from local cache.")
-  }
-  if  (!requireNamespace("readr", quietly = TRUE)) {
-    result <- utils::read.csv(vector_file, encoding = "UTF-8",stringsAsFactors = FALSE)
-    class(result) <- c("tbl_df", "tbl", "data.frame")
-  } else {
-    result <- readr::read_csv(vector_file)
-  }
-  dplyr::mutate(
+    cancensus.handle_status_code(response, NULL)
+    content <- httr::content(response, type = "text", encoding = "UTF-8")
+    result <- if (!requireNamespace("readr", quietly = TRUE)) {
+      dplyr::as_data_frame(utils::read.csv(content, stringsAsFactors = FALSE))
+    } else {
+      readr::read_csv(content)
+    }
+    result <- dplyr::mutate(
       result, type = factor(type),
       units = factor(units, levels = as.character(1:5),
                      labels = c("Number", "Percentage ratio (0.0-1.0)",
@@ -347,8 +353,22 @@ cancensus.list_vectors <- function(dataset, use_cache=TRUE) {
         grepl("^3.", add) ~ gsub(".", ", ", gsub("^3.", "Median of ", add),
                                  fixed = TRUE)
       )) %>%
-    dplyr::select(vector, type, label, units, parent_vector = parent,
-                  aggregation)
+      dplyr::select(vector, type, label, units, parent_vector = parent,
+                    aggregation)
+    attr(result, "last_updated") <- Sys.time()
+    save(result, file = cache_file)
+    result
+  } else {
+    message("Reading vector information from local cache.")
+    load(file = cache_file)
+    last_updated <- attr(result, "last_updated")
+    if (is.null(last_updated) ||
+          difftime(Sys.time(), last_updated, units = "days") > 1) {
+      warning(paste("Cached vectors list may be out of date. Set `use_cache =",
+                    "FALSE` to update it."))
+    }
+    result
+  }
 }
 
 #' Query the CensusMapper API for vectors with descriptions matching a searchterm.
@@ -449,14 +469,13 @@ cancensus.search_vectors <- function(searchterm, dataset) {
 #'
 #' @importFrom dplyr %>%
 cancensus.list_regions <- function(dataset, use_cache = TRUE) {
-  # TODO: Validate dataset?
   cache_dir <- system.file("cache/", package = "cancensus")
   cache_file <- paste0(cache_dir, dataset, "_regions.rda")
   if (!use_cache || !file.exists(cache_file)) {
     message("Querying CensusMapper API for regions data...")
     response <- httr::GET(paste0("https://censusmapper.ca/data_sets/", dataset,
                                  "/place_names.csv"))
-    cancensus.handle_status_code(response,  cache_file)
+    cancensus.handle_status_code(response, NULL)
     content <- httr::content(response, type = "text", encoding = "UTF-8")
     result <- if (!requireNamespace("readr", quietly = TRUE)) {
       dplyr::as_data_frame(utils::read.csv(content, stringsAsFactors = FALSE))
@@ -465,11 +484,18 @@ cancensus.list_regions <- function(dataset, use_cache = TRUE) {
     }
     result <- dplyr::select(result, region = geo_uid, name, level = type,
                             pop = population, municipal_status = flag)
+    attr(result, "last_updated") <- Sys.time()
     save(result, file = cache_file)
     result
   } else {
     message("Reading regions list from local cache.")
     load(file = cache_file)
+    last_updated <- attr(result, "last_updated")
+    if (is.null(last_updated) ||
+          difftime(Sys.time(), last_updated, units = "days") > 1) {
+      warning(paste("Cached regions list may be out of date. Set `use_cache =",
+                    "FALSE` to update it."))
+    }
     result
   }
 }
@@ -536,7 +562,9 @@ cancensus.labels <-  function(x) {
 cancensus.handle_status_code <- function(response,path){
   if (httr::status_code(response)!=200) {
     message=httr::content(response,as="text")
-    file.remove(path)
+    if (!is.null(path)) {
+      file.remove(path)
+    }
     if (httr::status_code(response)==401) {
       # Problem with API key
       stop(paste("Download of Census Data failed.",
